Allow multiple comment syntaxes for a file type
Assembly files (.S) might use // or /* */ comments. Indeed, the
binutils assembler doesn't support //-style comments, so we have to
allow /**/ as well.

This patch slightly generalises the format of COMMENT_CHARS to allow a
list of comment styles. It then iterates over the possible comment
styles to find a match.