Combine licence and CommentStyle objects early
This patch defines a new LicenceMatcher class, which wraps up a
CommentStyle and a LicenceHeader.

When the checker runs, the first thing it does is construct a
dictionary containing a LicenceMatcher object for each CommentStyle.
These are keyed by some opaque string type (the values now found in
COMMENT_CHARS).

The dictionary is passed to detect_comment_char (as the all_matchers
argument), and that function returns a list of matchers to be
used. (Before, it returned a list of CommentStyle objects to be used)

Finally, the check_file_with_matcher function (previously called
check_file_with_style) uses the matcher object to make sense of a
given file header.

Why do this rejig? One advantage is that it feels a bit cleaner (to
me). Rather than passing CommentStyle objects and licences around, we
now just have opaque LicenceMatcher objects.

The change also allows cleverer LicenceMatcher objects in the future:
there are plans afoot to allow regular expressions in the licence
field of the configuration. Rewriting like this means that the code
using a LicenceMatcher needn't care about the difference. It also
means that any sanity checking of the configuration file can happen at
"build time", rather than when we get to actually checking a file.