[bazel] use patched python rules and build wheels cache
This repository is brought in as a Bazel dependency in the OpenTitan
project. In order to support offline (air-gapped) Bazel builds of
OpenTitan, all dependencies must also build in an air-gapped
environment. Since this repository make use of the Bazel Python rules,
Bazel will try to fetch and install Python packages using pip, will not
not cache them in Bazel's repository cache (as Bazel has no knowledge of
pip).

Therefore a similar approach to lowRISC/opentitan:#12091 must be taken:
that is, a custom repository rule must be added to cache Python wheels,
so that Bazel will not attempt to fetch them in the airgapped
environment if they are already present.

This partially addresses lowRISC/opentitan:#12538.

Signed-off-by: Timothy Trippel <ttrippel@google.com>